requirements.txt
Flask
flask-socketio
pyserial 